+++ src/Client.php
@@ -10,6 +10,7 @@ use BeckonBilling\ApiClient\Resource\Articles;
 use BeckonBilling\ApiClient\Resource\Auth;
 use BeckonBilling\ApiClient\Resource\Customers;
 use BeckonBilling\ApiClient\Resource\DocumentTemplates;
+use BeckonBilling\ApiClient\Resource\Orders;
 use BeckonBilling\ApiClient\Resource\OutboundInvoices;
 use BeckonBilling\ApiClient\Resource\Quotes;
 use BeckonBilling\ApiClient\Resource\RecurringInvoices;
@@ -35,6 +36,7 @@ use BeckonBilling\ApiClient\Resource\Units;
  * @property-read Articles           $articles
  * @property-read Units              $units             Read-only.
  * @property-read DocumentTemplates  $documentTemplates Read-only.
+ * @property-read Orders             $orders
  * @property-read Quotes             $quotes
  * @property-read OutboundInvoices   $outboundInvoices
  * @property-read RecurringInvoices  $recurringInvoices
@@ -47,6 +49,7 @@ final class Client
     public readonly Articles $articles;
     public readonly Units $units;
     public readonly DocumentTemplates $documentTemplates;
+    public readonly Orders $orders;
     public readonly Quotes $quotes;
     public readonly OutboundInvoices $outboundInvoices;
     public readonly RecurringInvoices $recurringInvoices;
@@ -68,6 +71,7 @@ final class Client
         $this->articles = new Articles($this->transport);
         $this->units = new Units($this->transport);
         $this->documentTemplates = new DocumentTemplates($this->transport);
+        $this->orders = new Orders($this->transport);
         $this->quotes = new Quotes($this->transport);
         $this->outboundInvoices = new OutboundInvoices($this->transport);
         $this->recurringInvoices = new RecurringInvoices($this->transport);